+++ src/main/java/com/tristankechlo/toolleveling/commands/ToolLevelingCommand.java
@@ -6,10 +6,13 @@ import static net.minecraft.server.command.CommandManager.literal;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
+import com.tristankechlo.toolleveling.ToolLeveling;
 import com.tristankechlo.toolleveling.config.ConfigManager;
+import com.tristankechlo.toolleveling.network.ServerNetworkHandler;
 import com.tristankechlo.toolleveling.utils.Names;
 
 import net.minecraft.server.command.ServerCommandSource;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.ClickEvent;
 import net.minecraft.text.LiteralText;
 import net.minecraft.text.MutableText;
@@ -19,15 +22,15 @@ import net.minecraft.util.Formatting;
 public final class ToolLevelingCommand {
 
 	public static void register(CommandDispatcher<ServerCommandSource> dispatcher) {
-		LiteralArgumentBuilder<ServerCommandSource> toollevelingCommand = literal("toolleveling")
+		LiteralArgumentBuilder<ServerCommandSource> toollevelingCommand = literal("toolleveling").then(literal("config")
 				.requires((source) -> source.hasPermissionLevel(3))
-				.then(literal("config").then(literal("reload").executes(ToolLevelingCommand::configReload))
-						.then(literal("show")
-								.then(argument("identifier", EnumArgument.enumArgument(ConfigIdentifier.class))
-										.executes(ToolLevelingCommand::configShow)))
-						.then(literal("reset")
-								.then(argument("identifier", EnumArgument.enumArgument(ConfigIdentifier.class))
-										.executes(ToolLevelingCommand::configReset))));
+				.then(literal("reload").executes(ToolLevelingCommand::configReload))
+				.then(literal("show").then(argument("identifier", EnumArgument.enumArgument(ConfigIdentifier.class))
+						.executes(ToolLevelingCommand::configShow)))
+				.then(literal("reset").then(argument("identifier", EnumArgument.enumArgument(ConfigIdentifier.class))
+						.executes(ToolLevelingCommand::configReset))))
+				.then(literal("openitemvalues").requires((source) -> source.hasPermissionLevel(0))
+						.executes(ToolLevelingCommand::showScreen));
 		dispatcher.register(toollevelingCommand);
 	}
 
@@ -62,6 +65,17 @@ public final class ToolLevelingCommand {
 		}
 		ConfigManager.resetOneConfig(source.getServer(), identifier.id);
 		source.sendFeedback(new TranslatableText("commands.toolleveling.config.reset", identifier.id), true);
+		return 1;
+	}
+
+	private static int showScreen(CommandContext<ServerCommandSource> context) {
+		try {
+			ServerPlayerEntity player = context.getSource().getPlayer();
+			ServerNetworkHandler.sendOpenItemValues(player);
+		} catch (Exception e) {
+			ToolLeveling.LOGGER.error("Error while sending command '/openitemvalues'!\n" + e.getMessage());
+			return 0;
+		}
 		return 1;
 	}